Fill out 3 main sections, change image size, add hr
Entered relevant text in 'description", "ingredients" and "steps"
sections to provide more info on recipes. Now anyone can follow the
steps to make the recipes. Increased all image sizes - proportionally-
so that the width was more or less same as the width of the text below
it. Added a horizontal line for practice.